requirements.txt
requests==2.31.0
PyJWT==2.8.0
playsound==1.3.0